Fix preview from crashing.
